Implementing IdentityServer finished; Implementing communication between BatBetServiceAPI and Identity project through message broker (update user info after placing a bet);

+++ src/BatBetService/Controllers/BetsController.cs
@@ -1,6 +1,8 @@
 ﻿using BatBetDomain.DTOs.Request;
 using BatBetDomain.DTOs.Response;
 using BatBetDomain.Interfaces.Services;
+using BatBetServiceAPI.Extensions;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using System.Threading.Tasks;
 
@@ -24,18 +26,24 @@ namespace BatBetDomain.Controllers
             return Ok(await _betsService.GetBetById(id));
         }
 
+        [Authorize]
         [HttpPost]
-        public async Task<IActionResult> CreateBet(PlaceBetDto betDto, int userId)
+        public async Task<IActionResult> CreateBet(PlaceBetDto betDto)
         {
-            BetDto placedBet = await _betsService.CreateBet(betDto, userId);
+            BetDto placedBet = await _betsService
+                    .CreateBet(betDto, User.GetUserId(), User.GetUserClaims());
 
             return CreatedAtAction(nameof(GetBetById),
                 new { placedBet.Id }, placedBet);
         }
 
+        [Authorize]
         [HttpPut("{id}")]
         public async Task<IActionResult> UpdateBet(int id)
         {
+            //TODO: change this validation after configuring the identity user table and removing the current User table
+            if (string.IsNullOrEmpty(User.Identity.Name)) return Forbid();
+
             int result = await _betsService.UpdateBet(id);
 
             if (result != 0) return Ok($"Updated bet id: {result}");
@@ -43,9 +51,13 @@ namespace BatBetDomain.Controllers
             return BadRequest("Problem saving changes");
         }
 
+        [Authorize]
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteBet(int id)
         {
+            //TODO: change this validation after configuring the identity user table and removing the current User table
+            if (string.IsNullOrEmpty(User.Identity.Name)) return Forbid();
+
             bool result = await _betsService.DeleteBet(id);
 
             if (result) return NoContent();
